[#6] extracting metrics into a separate file
